+++ src/model/Model_Attachment.h
@@ -66,7 +66,7 @@ public:
     static wxString reftype_desc(const int RefTypeEnum);
 
     /** Return a dataset with attachments linked to a specific type*/
-    std::map<int, Data_Set> get_all(REFTYPE reftype);
+    std::map<int64, Data_Set> get_all(REFTYPE reftype);
 
     /** Return all attachments descriptions*/
     wxArrayString allDescriptions();
